Don't abort on missing config

+++ make-symlinks.sh
@@ -21,7 +21,7 @@ files="gitconfig config/terminator/config"
 ##########
 
 # create dotfiles_old in homedir
-echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
+echo -n "Creating $olddir for a backup of any existing dotfiles in ~ ..."
 mkdir -p $olddir
 echo "done"
 
@@ -34,13 +34,18 @@ echo "done"
 # symlinks from the homedir to any files in the ~/dotfiles directory specified
 # in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file $olddir
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if test -f "~/.$file"; then
+        echo "Moving any existing dotfile $file from ~ to $olddir"
+        mv ~/.$file $olddir
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/$file ~/.$file
+    fi
 done
 
 # Special handling
+echo "Modifying ~/.bashrc"
 echo "source $dir/bashrc" >> ~/.bashrc
+echo "Modifying ~/.profile"
 echo "source $dir/profile" >> ~/.profile
+echo "Modifying ~/.inputrc"
 echo "\$include $dir/inputrc" >> ~/.inputrc
